+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod reader;
-pub mod tile;
 pub mod world;
 pub mod writer;
